requirements.txt
django_environ==0.4.5
Jinja2==2.10.1
colorama==0.4.1
ordereddict==1.1
six==1.12.0
paramiko==2.7.1
environ==1.0
GitPython==3.1.0
PyYAML==5.3
